+++ changelog.md
@@ -4,6 +4,7 @@
 - Fixed TextBox inputs not allowing user to set an decimal point on certain actions
 - Added **Set Gravity**,**Speed Player** & **Noclip** Events
 - Added a delete command comfirmation popup
+- Fixed the texture scaling on the action & event nodes
 # v0.1.5-beta.1
 - Added **Sound Effect** & **Stop All Sounds** Events
 - Changed the **Wait** Action value to use float value instead of just integer.